Putting secular.py in a presentable form
I put the guts of the code inside functions and finished the command line
wrapper.  I switched from my homebrew constants module to the astropy
constants module.  I got rid of the REO prescription.

The code still needs to be debugged and needs to undergo extensive testing.
There are still major structural changes that are required, so the code is
in nowhere near a publishable form yet.