Turned evolution function into generator
The function that now evolves a hierarchical triple in time is now a
generator.  This makes the module much more easily extensible.  Minor
modifications to the output can be handled by different wrapper functions
rather than altering the heart of the evolution code.

Additionally, the code now outputs the initial conditions in JSON format.