Added semi-analytic calculation of the period
This function semi-analytically calculates the time between flips (or
otherwise the period of oscillations in j_z or Omega if the triple doesn't
flip.)  The period calculation for non-flipping systems may not be accurate;
this functionality has not yet been tested.

I additionally added a function to calculate whether the triple flips.