Added flip_time function to ts_vector
The flip_time function prints out only the times during which a triple
flips.  This makes it easy to calculate the period of the EKM using the
secular code.

I also added a function which resets the triple back to time zero so that it
can be integrated multiple times.  This does not yet work properly because
the integrator itself is not reset, only the variables in the class.