Added sec. flip period, changed analytic p. calc
I added a method to Triple_vector to calculate the period of flips.

I also changed the analytic calculation of the flip period.  Before it was
the full period of EKM oscillations.  However, the time between flips is
what we are generally interested in, and this is one-half the full period.
I therefore changed the calculated period to half its original definition.
Note that this brings the calculation in line with what is in the paper.