Improved limit calculation in period method
The calculation of the limits of the integral in the period method did not
work over the entire range of parameter space.  In particular, it would not
work if the flips occured for Omega = pi (or maybe 0).  Now the limit
calculation checks which value of Omega flips occur at and does the limit
calculation accordingly.