Fixing up package structure
The code now more or less works and is written in an object-oriented style.
There are still a number of major improvements that need to be made which
are listed in the TODO list.  I cannot guarantee that the code all works as
advertised at the moment because the test suite is still just getting
started.  But the module as it currently exists works for basic numerical
integration tasks.

- Changed kl_period so that the functions accept triple objects
- Added a variety of tests.  (More still need to be added!)
- Added code to ignore DeprecationWarnings to __init__.py
- Added items to the TODO list