Fixed collision bug in delaunay
There was a bug in TripleDelaunay where the units for collision testing were
not handled correctly.  This led to the code stating that certain triples
had undergone collisions, when the actually hadn't.  This has been fixed and
a test has been added.

I also added a setup.cfg file.  The setup.py file has been updated with a
new version number.